+++ src/video_core/video_core.cpp
@@ -25,10 +25,10 @@ std::function<void()> g_screenshot_complete_callback;
 Layout::FramebufferLayout g_screenshot_framebuffer_layout;
 
 /// Initialize the video core
-Core::System::ResultStatus Init() {
+Core::System::ResultStatus Init(Frontend& frontend) {
     Pica::Init();
 
-    g_renderer = std::make_unique<Renderer>();
+    g_renderer = std::make_unique<Renderer>(frontend);
     Core::System::ResultStatus result{g_renderer->Init()};
 
     if (result != Core::System::ResultStatus::Success) {
@@ -62,8 +62,9 @@ void RequestScreenshot(void* data, std::function<void()> callback,
 }
 
 u16 GetResolutionScaleFactor() {
-    return !Settings::values.resolution_factor ? Frontend::GetFramebufferLayout().GetScalingRatio()
-                                               : Settings::values.resolution_factor;
+    return !Settings::values.resolution_factor
+               ? Core::System::GetInstance().GetFrontend().GetFramebufferLayout().GetScalingRatio()
+               : Settings::values.resolution_factor;
 }
 
 } // namespace VideoCore
