+++ src/video_core/video_core.h
@@ -10,6 +10,7 @@
 #include "core/frontend.h"
 
 class Renderer;
+class Frontend;
 
 namespace VideoCore {
 
@@ -29,7 +30,7 @@ extern std::function<void()> g_screenshot_complete_callback;
 extern Layout::FramebufferLayout g_screenshot_framebuffer_layout;
 
 /// Initialize the video core
-Core::System::ResultStatus Init();
+Core::System::ResultStatus Init(Frontend& frontend);
 
 /// Shutdown the video core
 void Shutdown();
